refactor(init): split init flows into separate functions (#753)
