+++ packages/@sanity/cli/src/actions/init/initApp.ts
@@ -1,99 +1,78 @@
 import {styleText} from 'node:util'
 
-import {type Output, type TelemetryUserProperties} from '@sanity/cli-core'
+import {subdebug, type TelemetryUserProperties} from '@sanity/cli-core'
 import {logSymbols} from '@sanity/cli-core/ux'
 import {type TelemetryTrace} from '@sanity/telemetry'
 
 import {type InitStepResult} from '../../telemetry/init.telemetry.js'
 import {type EditorName} from '../mcp/editorConfigs.js'
+import {InitError} from './initError.js'
 import {getPostInitMCPPrompt} from './initHelpers.js'
 import {type RepoInfo} from './remoteTemplate.js'
 import {scaffoldAndInstall, selectTemplate} from './scaffoldTemplate.js'
+import {type InitContext, type InitOptions} from './types.js'
+
+const debug = subdebug('init:app')
+
+interface InitAppParams {
+  defaults: {projectName: string}
+  mcpConfigured: EditorName[]
+  options: InitOptions
+  organizationId: string | undefined
+  output: InitContext['output']
+  outputPath: string
+  remoteTemplateInfo: RepoInfo | undefined
+  sluggedName: string
+  trace: TelemetryTrace<TelemetryUserProperties, InitStepResult>
+  workDir: string
+}
 
 export async function initApp({
-  autoUpdates,
   defaults,
-  error,
-  git,
-  noGit,
   mcpConfigured,
+  options,
   organizationId,
   output,
   outputPath,
-  overwriteFiles,
-  packageManager,
   remoteTemplateInfo,
   sluggedName,
-  template,
-  templateToken,
   trace,
-  typescript,
-  unattended,
   workDir,
-}: {
-  autoUpdates: boolean
-  defaults: {projectName: string}
-  error: Output['error']
-  git?: boolean | string
-  noGit?: boolean
-  mcpConfigured: EditorName[]
-  organizationId: string | undefined
-  output: Output
-  outputPath: string
-  overwriteFiles?: boolean
-  packageManager?: string
-  remoteTemplateInfo: RepoInfo | undefined
-  sluggedName: string
-  template?: string
-  templateToken?: string
-  trace: TelemetryTrace<TelemetryUserProperties, InitStepResult>
-  typescript?: boolean
-  unattended: boolean
-  workDir: string
-}): Promise<void> {
-  const {
-    template: resolvedTemplate,
-    templateName,
-    useTypeScript,
-  } = await selectTemplate({
-    remoteTemplateInfo,
-    template,
-    trace,
-    typescript,
-    unattended,
-  })
+}: InitAppParams): Promise<void> {
+  debug('Scaffolding app template')
 
-  if (!remoteTemplateInfo && !resolvedTemplate) {
-    error(`Template "${templateName}" not found`, {exit: 1})
+  // Prompt for template and TypeScript
+  const {template, templateName, useTypeScript} = await selectTemplate(
+    options,
+    remoteTemplateInfo,
+    trace,
+  )
+  if (!remoteTemplateInfo && !template) {
+    throw new InitError(`Template "${templateName}" not found`, 1)
   }
 
-  await scaffoldAndInstall({
-    autoUpdates,
+  // Bootstrap, install deps, git init
+  const {pkgManager} = await scaffoldAndInstall({
     datasetName: '',
     defaults,
     displayName: '',
-    git,
-    noGit,
+    options,
     organizationId,
     output,
     outputPath,
-    overwriteFiles,
-    packageManager,
     projectId: '',
     remoteTemplateInfo,
     sluggedName,
     templateName,
-    templateToken,
     trace,
-    unattended,
     useTypeScript,
     workDir,
   })
 
-  const isCurrentDir = outputPath === process.cwd()
+  // App-specific success messages
+  const isCurrentDir = outputPath === workDir
   const goToProjectDir = `\n(${styleText('cyan', `cd ${outputPath}`)} to navigate to your new project directory)`
 
-  //output for custom apps here
   output.log(
     `${logSymbols.success} ${styleText(['green', 'bold'], 'Success!')} Your custom app has been scaffolded.`,
   )
@@ -118,4 +97,6 @@ export async function initApp({
   output.log(`npx sanity docs browse     to open the documentation in a browser`)
   output.log(`npx sanity dev             to start the development server for your app`)
   output.log(`npx sanity deploy          to deploy your app`)
+
+  debug('App scaffolding complete (pkgManager=%s)', pkgManager)
 }
