+++ packages/@sanity/cli/src/telemetry/init.telemetry.ts
@@ -31,6 +31,7 @@ interface CreateOrSelectDatasetStep {
   datasetName: string
   selectedOption: 'create' | 'none' | 'select'
   step: 'createOrSelectDataset'
+
   visibility?: 'private' | 'public'
 }
